Fixes for KittyCommand

+++ CoreCodedChatbot.Library/Models/Data/RedditRandomJsonModel.cs
@@ -76,7 +76,7 @@ namespace CoreCodedChatbot.Library.Models.Data
         public object content_categories { get; set; }
         public bool is_self { get; set; }
         public object mod_note { get; set; }
-        public int created { get; set; }
+        public float created { get; set; }
         public string link_flair_type { get; set; }
         public int wls { get; set; }
         public object banned_by { get; set; }
@@ -121,7 +121,7 @@ namespace CoreCodedChatbot.Library.Models.Data
         public bool stickied { get; set; }
         public string url { get; set; }
         public int subreddit_subscribers { get; set; }
-        public int created_utc { get; set; }
+        public float created_utc { get; set; }
         public Media media { get; set; }
         public bool is_video { get; set; }
         public string post_hint { get; set; }
